pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autopilot-video"
version = "0.1.0"
description = "Automated video editing pipeline — from raw footage to published video"
license = "AGPL-3.0-or-later"
requires-python = ">=3.11"
dependencies = [
    # Core
    "click",
    "numpy",
    "Pillow",
    "pyyaml",
    "requests",
    # ML — speech & audio
    "whisperx",
    "faster-whisper",
    "panns-inference",
    # ML — vision
    "ultralytics",
    "insightface",
    "transformers",
    "qwen-vl-utils",
    "faiss-gpu",
    "scikit-learn",
    # ML — generation
    "audiocraft",
    "kokoro",
    # Video
    "moviepy>=2.0",
    "pyav",
    "opentimelineio",
    "scenedetect",
    # APIs
    "anthropic",
    "google-api-python-client",
]

[project.scripts]
autopilot = "autopilot.cli:main"

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "pyright",
]
vllm = [
    "vllm",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.coverage.run]
source = ["autopilot"]
omit = ["*/tests/*"]

[tool.coverage.report]
show_missing = true
fail_under = 70

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
reportMissingImports = false
